% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdamDB-class.R
\docType{class}
\name{AdamDB-class}
\alias{AdamDB-class}
\alias{.AdamDB}
\alias{AdamDB}
\title{\code{AdamDB}}
\usage{
AdamDB(object)
}
\arguments{
\item{object}{(\code{list} of \code{data.frame}) to convert to \code{\link{AdamDB}}.}
}
\description{
The \code{AdamDB} class is an extension of \code{list} with additional validation criteria.
}
\details{
The additional criteria are:
\itemize{
\item The list elements must be of class \code{data.frame}.
}
}
\examples{
library(scda.2021)
dat <- rcd_2021_03_22
ad <- AdamDB(dat)
}
