% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{append_rules}
\alias{append_rules}
\title{Append rules to an existing rules sets}
\usage{
append_rules(rules_base, rule_add)
}
\arguments{
\item{rules_base}{(\code{rules}) to update.}

\item{rule_add}{(\code{rules}) to add/update.}
}
\description{
Append rules to an existing rules sets
}
