% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{apply_reformat}
\alias{apply_reformat}
\title{Reformat values}
\usage{
apply_reformat(db, format = NULL)
}
\arguments{
\item{db}{(\code{dm}  or \code{list} of \code{data.frame}) object input.}

\item{format}{(\code{list}) in a specific format.}
}
\value{
a \code{dm} object. If not reformatted, value remain unchanged.
}
\description{
Reformat values
}
\note{
Using the keyword \code{All} as a table name will change the corresponding variable in every table where it appears.
}
\keyword{internal}
