% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{as.list.rule}
\alias{as.list.rule}
\title{Convert Rule to List}
\usage{
\method{as.list}{rule}(x, ...)
}
\arguments{
\item{x}{(\code{rule}) to convert.}

\item{...}{not used.}
}
\description{
Convert Rule to List
}
\examples{

x <- rule("a" = c("a", "b"), "X" = "x", .to_NA = c("v", "w"))
as.list(x)
}
