% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_adsub_adsl.R
\name{assert_names_collision}
\alias{assert_names_collision}
\title{Assert whether any two names are identical.}
\usage{
assert_names_collision(vars_nam)
}
\description{
Assert whether any two names are identical.
}
\keyword{internal}
