% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_adsub_adsl.R
\name{assert_names_notadsl}
\alias{assert_names_notadsl}
\title{Assert that names are not in adsl}
\usage{
assert_names_notadsl(vars_nam, df)
}
\description{
Assert that names are not in adsl
}
\keyword{internal}
