% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{assert_reformat}
\alias{assert_reformat}
\title{Assert the Reformatting Map.}
\usage{
assert_reformat(map)
}
\arguments{
\item{map}{(\code{list})}
}
\value{
\code{NULL} if the \code{map} object fits the criteria for a mapping list.
}
\description{
Assert the Reformatting Map.
}
\examples{
my_map <- list(
  df1 = list(
    char = list(
      "A" = c("a", "k"),
      "B" = "b"
    ),
    char2 = list(
      "A" = c("a", "k"),
      "B" = "b"
    )
  ),
  df2 = list(
    num = list(
      "11" = "1",
      "22" = NA
    )
  ),
  All = list(
    fact = list(
      "F1" = "f1",
      "F2" = "f2"
    ),
    other = list(
      "x" = "X"
    )
  )
)

assert_reformat(my_map)

my_map <- list(
  df0 = NULL,
  df1 = list(
    char = NULL,
    char2 = list(
      "A" = c("a", "k"),
      "B" = "b"
    )
  )
)

assert_reformat(my_map)
}
