% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_valid_format}
\alias{assert_valid_format}
\title{Assert List can be used as Format.}
\usage{
assert_valid_format(object)
}
\arguments{
\item{object}{(\code{list}) to assert.}
}
\value{
invisible \code{TRUE} or an error message if the criteria are not fulfilled.
}
\description{
Assert List can be used as Format.
}
\examples{
format <- list(
  df1 = list(
    var1 = list("X" = "x", "N" = c(NA, ""))
  ),
  df2 = list(
    var1 = list(),
    var2 = list("f11" = "F11", "NN" = NA)
  ),
  df3 = list()
)

assert_valid_format(format)
}
