% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{attr_label}
\alias{attr_label}
\title{Setting the Label Attribute}
\usage{
attr_label(var, label)
}
\arguments{
\item{var}{(\code{object}) whose label attribute can be set.}

\item{label}{(\code{character}) the label to add.}
}
\value{
\code{object} with label attribute.
}
\description{
Setting the Label Attribute
}
\examples{
x <- c(1:10)
attr(x, "label")

y <- attr_label(x, "my_label")
attr(y, "label")
}
