% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{combineListRules}
\alias{combineListRules}
\title{Combine Rules Found in Lists of Rules.}
\usage{
combineListRules(x, val, ...)
}
\arguments{
\item{x}{(\code{list}) of \code{rule} objects.}

\item{val}{(\code{list}) of \code{rule} objects.}

\item{...}{passed to \code{combine_rules}.}
}
\value{
a \code{list} of \code{rule} objects.
}
\description{
Combine Rules Found in Lists of Rules.
}
\examples{
l1 <- list(
  r1 = rule(
    "first" = c("from ori rule", "FROM ORI RULE"),
    "last" = c(NA, "last")
  ),
  r2 = rule(
    ANYTHING = "anything"
  )
)

l2 <- list(
  r1 = rule(
    "first" = c("F", "f"),
    "second" = c("S", "s"),
    "third" = c("T", "t"),
    .to_NA = "something"
  ),
  r3 = rule(
    SOMETHING = "something"
  )
)

combineListRules(l1, l2)
}
