% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreorder_levels.R
\name{coreorder_levels}
\alias{coreorder_levels}
\title{Reorder Two Columns Levels Simultaneously}
\usage{
coreorder_levels(df, primary, secondary, levels_primary)
}
\arguments{
\item{df}{(\code{data.frame}) with PARAM and PARAMCD variables.}

\item{paramcd_levels}{(\code{character}) ordered levels of PARAMCD.}
}
\description{
Reorder Two Columns Levels Simultaneously
}
\examples{

df <- data.frame(SUBJID = 1:3, PARAMCD = factor(c("A", "B", "C")), PARAM = factor(paste("letter", LETTERS[1:3])))
coreorder_levels(df, "PARAMCD", "PARAM", levels_primary = c("C", "A", "B"))

}
