% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunlin-package.R
\docType{package}
\name{dunlin-package}
\alias{dunlin}
\alias{dunlin-package}
\title{\code{Dunlin} Package}
\description{
A collection of functions to preprocess data and organize them in a format amenable to use by chevron.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://insightsengineering.github.io/dunlin/}
  \item \url{https://github.com/insightsengineering/dunlin/}
  \item Report bugs at \url{https://github.com/insightsengineering/dunlin/issues}
}

}
\author{
\strong{Maintainer}: Liming Li \email{liming.li@roche.com}

Authors:
\itemize{
  \item Benoit Falquet \email{benoit.falquet@roche.com}
}

Other contributors:
\itemize{
  \item Xiaoli Duan \email{xiaoli.duan@roche.com} [contributor]
  \item Pawel Rucki \email{pawel.rucki@roche.com} [contributor]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
