% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_as_factor}
\alias{h_as_factor}
\title{Transforming Empty Strings and White Spaces to Explicit NAs while Preserving Label}
\usage{
h_as_factor(x, na_level = "<Missing>")
}
\arguments{
\item{x}{(\code{character} or \code{factor}) input to be turned into factor with explicit missing level.}

\item{na_level}{(\code{character}) the label to encode missing levels.}
}
\value{
\code{factor} with explicit NA and the same label as the input.
}
\description{
Transforming Empty Strings and White Spaces to Explicit NAs while Preserving Label
}
\details{
This function preserves the label attribute.
}
\examples{
char1 = c(" ", "    ", "a", "b", "", "", NA)
attr(char1, "label") <- "my_label"

h_as_factor(char1)

}
