% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explicit_na.R
\name{h_df_explicit}
\alias{h_df_explicit}
\title{Encode Categorical Missing Values in a \code{data.frame}.}
\usage{
h_df_explicit(
  df,
  omit_columns = NULL,
  char_as_factor = TRUE,
  na_level = "<Missing>"
)
}
\arguments{
\item{omit_columns}{(\code{character}) the names of the columns to omit from processing.}

\item{char_as_factor}{(\code{logical}) should character columns be converted into factor.}

\item{na_level}{(\code{string}) the label to encode missing levels.}
}
\value{
a \code{data.frame} object with explicit missing levels.
}
\description{
Encode Categorical Missing Values in a \code{data.frame}.
}
\keyword{internal}
