% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{h_expand_all_datasets}
\alias{h_expand_all_datasets}
\title{Propagate the rules for all datasets}
\usage{
h_expand_all_datasets(format_list, ls_datasets = NULL)
}
\arguments{
\item{ls_datasets}{(\code{character}) the name of all datasets in the object to reformat.}
}
\value{
a nested \code{list} attributing a rule to be applied to specific variables of specific datasets.
}
\description{
Propagate the rules for all datasets
}
\details{
the rules described  under \code{all_datasets} are propagated to all data sets for the corresponding variables
except in data sets where a rule is already attributed to the same variable.
}
\keyword{internal}
