% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_unnest_format}
\alias{h_unnest_format}
\title{Unnest a list of mappings}
\usage{
h_unnest_format(.lst)
}
\arguments{
\item{.lst}{(\code{list}) of mapping.}
}
\value{
\verb{named character}.
}
\description{
Unnest a list of mappings
}
\examples{

l <- list(A = "A", B = c("X", "Y"), Missing = c(NA))
h_unnest_format(l)
}
