% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_ws_to_explicit_na}
\alias{h_ws_to_explicit_na}
\title{Transforming Empty Strings and White Spaces to Explicit NAs}
\usage{
h_ws_to_explicit_na(x, na_level = "<Missing>")
}
\arguments{
\item{x}{(\code{character} or \code{factor}) where empty of white space should be transformed to NAs.}
}
\value{
\code{factor} with explicit NA
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

SAS imports missing data as empty strings or white spaces. This helper function is a thin wrapper around
\link{h_ws_to_na} which replaces them with explicit missing level.
}
\examples{

char1 = c(" ", "    ", "a", "b", "", "")
h_ws_to_explicit_na(char1)

fact1 <- as.factor(char1)
h_ws_to_explicit_na(fact1)

}
