% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_adsub_adsl.R
\name{join_adsub_adsl}
\alias{join_adsub_adsl}
\title{Join \code{adsub} to \code{adsl}}
\usage{
join_adsub_adsl(
  adam_db,
  keys = c("USUBJID", "STUDYID"),
  continuous_var = "all",
  categorial_var = "all",
  continuous_suffix = "",
  categorial_suffix = "_cat"
)
}
\arguments{
\item{keys}{(\code{character}) the name of the columns in \code{adsl} uniquely identifying a row.}

\item{continuous_var}{(\code{character}) the value of a parameter in the \code{PARAMCD} column of the \code{adsub} table from which
columns containing continuous values should be created.}

\item{categorial_var}{(\code{character}) the value of a parameter in the \code{PARAMCD} column of the \code{adsub} table from which
columns containing categorical values should be created.}

\item{continuous_suffix}{(\code{string}) the suffixes to add to the newly generated columns containing continuous values.}

\item{categorial_suffix}{(\code{string}) the suffixes to add to the newly generated columns containing categorical values.}

\item{db}{(\code{dm}) object input with an \code{adsl} and \code{adsub} table.}
}
\value{
a \code{dm} object with new columns in the \code{adsl} table.
}
\description{
Join \code{adsub} to \code{adsl}
}
\examples{
adsl <- data.frame(
  USUBJID = c("S1", "S2", "S3", "S4"),
  STUDYID = "My_study",
  AGE = c(60, 44, 23, 31)
)

adsub <- data.frame(
  USUBJID = c("S1", "S2", "S3", "S4", "S1", "S2", "S3"),
  STUDYID = "My_study",
  PARAM = c("weight", "weight", "weight", "weight", "height", "height", "height"),
  PARAMCD = c("w", "w", "w", "w", "h", "h", "h"),
  AVAL = c(98, 75, 70, 71, 182, 155, 152),
  AVALC = c(">80", "<=80", "<=80", "<=80", ">180", "<=180", "<=180")
)

db <- dm::dm(adsl, adsub)

db <- dm_add_pk(db, adsl, STUDYID)
db <- dm_add_fk(db, adsub, STUDYID, adsl)

x <- join_adsub_adsl(adam_db = db)
x <- join_adsub_adsl(adam_db = db, continuous_var = c("w", "h"), categorial_var = "h")
}
