% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{log_filter}
\alias{log_filter}
\title{Filter Data with Log}
\usage{
log_filter(data, ...)
}
\arguments{
\item{data}{(\code{data.frame}) input data to subset, or named (\code{list}) of (\code{data.frame}).}

\item{table}{(\code{string}) table name.}

\item{condition}{(\code{expression}) subset condition. Must eval as logical.}
}
\description{
Filter Data with Log
}
\details{
\code{log_filter}
}
