% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot.R
\name{mini_pivot_wider}
\alias{mini_pivot_wider}
\title{Transforming data.frame into Wider Format}
\usage{
mini_pivot_wider(data, obs_id, param_from, value_from)
}
\arguments{
\item{data}{(\code{data.frame}) to be pivoted.}

\item{obs_id}{(\code{character}) the name of the column identifying the observations. It will correspond to the row names
of the output.}

\item{param_from}{(\code{character}) the name of the column containing the names of the parameters to be pivoted. The
unique values in this column will become column names in the output.}

\item{value_from}{(\code{character}) the name of the column containing the values that will populate the output.}
}
\value{
\code{data.frame} in a wide format.
}
\description{
Transforming data.frame into Wider Format
}
\details{
instead of nesting duplicated values, the function will throw an error if the same same parameter is
provided twice for the same observation.
}
\examples{
test_data <- data.frame(
  the_obs = c("A", "A", "A", "B", "B", "B", "C", "D"),
  the_obs2 = c("Ax", "Ax", "Ax", "Bx", "Bx", "Bx", "Cx", "Dx"),
  the_param = c("weight", "height", "gender", "weight", "gender", "height", "height", "precondition"),
  the_val = c( 65, 165, "M", 66, "F", 166, 155, TRUE)
 )

mini_pivot_wider(test_data, "the_obs", "the_param", "the_val" )

}
