% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modify_rule_attr}
\alias{modify_rule_attr}
\title{Extract Rule Parameter from reformat argument}
\usage{
modify_rule_attr(..., format)
}
\arguments{
\item{...}{argument to be passed to the rule.}

\item{format}{(\code{rule}) whose default attributes have to be modified.}
}
\description{
Extract Rule Parameter from reformat argument
}
\note{
NULL can be passed as argument value.
}
\examples{
format <- rule("X" = "x", "Y" = "y", .drop = FALSE)
modify_rule_attr(.drop = TRUE, not_att = "--", .to_NA = c(" "), format = format)

}
\keyword{internal}
