% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{modifyListRule}
\alias{modifyListRule}
\title{Combine Rules Found in Lists of Rules.}
\usage{
modifyListRule(x, val)
}
\description{
Combine Rules Found in Lists of Rules.
}
\examples{
l1 <- list(
  r1 = rule(
    "first" = c("from ori rule", "FROM ORI RULE"),
    "last" = c(NA, "last")
  ),
  r2 = rule(
    ANYTHING = "anything"
  )
)

l2 <- list(
  r1 = rule(
    "first" = c("F", "f"),
    "second" = c("S", "s"),
    "third" = c("T", "t"),
    .to_NA = "something"
  ),
  r3 = rule(
    SOMETHING = "something"
  )
)

modifyListRule(l1, l2)



}
