% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{print_log}
\alias{print_log}
\alias{print_log.data.frame}
\alias{print_log.list}
\title{Print Log}
\usage{
print_log(data, incl)

\method{print_log}{data.frame}(data, incl = TRUE)

\method{print_log}{list}(data, incl = TRUE)
}
\arguments{
\item{data}{(\code{list} of \code{data.frame} or \code{data.frame}) filtered with \code{log_filter}.}

\item{incl}{(\code{flag}) should information about unfiltered \code{data.frame} be printed.}
}
\description{
Print Log
}
\examples{
data <- log_filter(iris, Sepal.Length >= 7)
print_log(data)
data <- log_filter(list(adsl = iris, iris2 = iris), Sepal.Length >= 7, "adsl", character(0), "my filter")
print_log(data)
}
