% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagate.R
\name{propagate}
\alias{propagate}
\title{Propagate Column}
\usage{
propagate(db, from, add, by, safe = FALSE)
}
\arguments{
\item{db}{(\code{dm}) object for which some variable need to be propagated.}

\item{from}{(\code{string}) the name of the table where the variables to propagate are stored.}

\item{add}{(\code{character}) the names of the variables to propagate.}

\item{by}{(\code{character}) the key binding the \code{from} table to the other tables.}

\item{safe}{(\code{flag}) should the key be checked for uniqueness in the \code{from} table.}
}
\value{
updates \code{dm} object.
}
\description{
\code{propagate}copy columns from a given table of a \code{dm} object to all tables based on other common columns. If several
rows are associated with the same key, the rows will be duplicated in the receiving tables. In safe mode, the key
must be unique in the original table.
}
\examples{
db <- dm::dm_nycflights13()
res <- propagate(db, "airlines", "name", "carrier", safe = TRUE)

}
