% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{read_rules}
\alias{read_rules}
\title{Read yaml File describing \code{rule}}
\usage{
read_rules(file)
}
\arguments{
\item{file}{(\code{string}) of path to the rule yaml file.}
}
\description{
Read yaml File describing \code{rule}
}
