% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remap.R
\name{remap}
\alias{remap}
\title{Remap values}
\usage{
remap(db, map)
}
\arguments{
\item{db}{(\code{dm}) object input.}

\item{map}{(\code{list}) in a specific format.}
}
\value{
a \code{dm} object with re coded variables as factor.
}
\description{
Remap values
}
\examples{

df1 <- data.frame(
  "char" = c("a", "b", NA, "a", "k", "x"),
  "fact" = factor(c("f1", "f2", NA, NA, "f1", "f1")),
  "logi" = c(NA, FALSE, TRUE, NA, FALSE, NA)
)
df2 <- data.frame(
  "char" = c("a", "b", NA, "a", "k", "x"),
  "fact" = factor(c("f1", "f2", NA, NA, "f1", "f1")),
  "num" = 1:6
)

db <- dm(df1, df2)

my_map <- list(
  df1 = list(
    char = list(
      "A" = c("a", "k"), 
      "B" = "b"
    ),
    fact = list(
      "F1" = "f1",
      "F2" = "f2"
    )
  ),
  df2 = list(
    num = list(
      "11" = 1,
      "22" = 2
    )
  )
)

res <- remap(db, my_map)


}
