% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_safe.R
\name{remove_whisker}
\alias{remove_whisker}
\title{Remove whisker values}
\usage{
remove_whisker(x)
}
\arguments{
\item{x}{Named (\code{character}) input.}
}
\value{
invisible \code{NULL}. Removes \code{x} from the whisker environment.
}
\description{
Remove whisker values
}
