% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_safe.R
\name{render_safe}
\alias{render_safe}
\title{Render whiskers safely}
\usage{
render_safe(x)
}
\arguments{
\item{x}{(\code{character}) input to be rendered safely.}

\item{envir}{(\code{environment}) in which input is rendered.}
}
\description{
Render whiskers safely
}
