% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{rules}
\alias{rules}
\title{Create rule sets}
\usage{
rules(..., .lst = list(...))
}
\arguments{
\item{...}{Named arguments of rules.}

\item{.lst}{(\code{list}) of rules.}
}
\description{
Create rule sets
}
