% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_safe.R
\name{safe_transformer}
\alias{safe_transformer}
\title{Safe transformer}
\usage{
safe_transformer(text, envir)
}
\description{
Safe transformer
}
\details{
Obtain content in global environment by default.
If not found, use the environment here.
}
\keyword{internal}
