% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdamDB-validate.R
\name{validate_AdamDB}
\alias{validate_AdamDB}
\alias{validate_tables}
\alias{validate_primary_key}
\title{Internal Helper Functions for Validation of \code{AdamDB} Objects}
\usage{
validate_tables(object, criteria)

validate_primary_key(object, primary_table, primary_key)
}
\arguments{
\item{object}{(\code{list} of \code{data.frame}) to check for validity.}

\item{criteria}{(\code{list}) with the name of the mandatory tables and their required columns.}

\item{primary_table}{(\code{string}) the name of the table used as primary table.}

\item{primary_key}{(\code{string}) the name of the column used as key.}
}
\value{
A character vector with the validation failure messages,
or \code{NULL} in case validation passes.
}
\description{
These functions are only used internally to validate the format of an input \code{list} object and therefore
not exported.
}
\section{Functions}{
\itemize{
\item \code{validate_tables}: validate the presence of mandatory tables and columns.

\item \code{validate_primary_key}: validate that key values are present in the primary table and not duplicated.
}}

