% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{write_format}
\alias{write_format}
\title{Write Format yaml file}
\usage{
write_format(format, file)
}
\arguments{
\item{format}{(\code{list}) object.}

\item{file}{(\code{string}) file path to write to.}
}
\description{
Write Format yaml file
}
