% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{write_rules}
\alias{write_rules}
\title{Write rule}
\usage{
write_rules(rules, file, append = TRUE)
}
\arguments{
\item{rules}{(\code{rules}) object.}

\item{file}{(\code{string}) file path to write to.}

\item{append}{(\code{flag}) indicator of whether to overwrite or add rules.}
}
\description{
Write rule
}
