% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_safe.R
\name{add_whisker}
\alias{add_whisker}
\title{Add whisker values}
\usage{
add_whisker(x)
}
\arguments{
\item{x}{Named (\code{character}) input.}
}
\value{
invisible \code{NULL}. Assign the key-value pair provided as argument in the whisker environment.
}
\description{
Add whisker values
}
\details{
The names of the character gives the string to be replaced and the value gives the new string.
}
\examples{
my_whiskers <- c(Placeholder = "Replacement", Placeholder2 = "Replacement2")
add_whisker(my_whiskers)
}
