% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{apply_reformat}
\alias{apply_reformat}
\title{Reformat values}
\usage{
apply_reformat(db, format = NULL)
}
\arguments{
\item{db}{(\code{dm}) object input.}

\item{format}{(\code{list}) in a specific format.}
}
\value{
a \code{dm} object. If not reformatted, value remain unchanged.
}
\description{
Reformat values
}
\note{
Using the keyword \code{All} as a table name will change the corresponding variable in every table where it appears.
}
\examples{
library(dm)

df1 <- data.frame(
  "char" = c("a", "b", NA, "a", "k", "x"),
  "fact" = factor(c("f1", "f2", NA, NA, "f1", "f1"))
)
df2 <- data.frame(
  "char" = c("a", "b", NA, "a", "k", "x"),
  "fact" = factor(c("f1", "f2", NA, NA, "f1", "f1"), levels = c("f1", "f2", "fx"))
)

db <- dm(df1, df2)

new_format <- list(
  df1 = list(
    char = rule(
      "A" = c("a", "k"),
      "B" = "b"
    )
  ),
  ALL = list(
    fact = rule(
      "F1" = "f1",
      "F2" = "f2",
      "<Missing>" = NA
    ),
    other = rule(
      "x" = "X"
    )
  )
)

res <- apply_reformat(db, new_format)
}
