% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunlin-package.R
\docType{package}
\name{dunlin-package}
\alias{dunlin}
\alias{dunlin-package}
\title{Dunlin Package}
\description{
A collection of functions to preprocess data and organize them in a format amenable to use by chevron.
}
\author{
\strong{Maintainer}: NEST \email{basel.nestcicd@roche.com}

Authors:
\itemize{
  \item Heng Wang
  \item Benoit Falquet \email{benoit.falquet@roche.com}
}

Other contributors:
\itemize{
  \item Roche [copyright holder, funder]
  \item Tim Barnett \email{timothy.barnett@roche.com} [contributor]
}

}
