% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunlin-package.R
\docType{package}
\name{dunlin-package}
\alias{dunlin}
\alias{dunlin-package}
\title{Dunlin Package}
\description{
A collection of functions to preprocess data and organize them in a format amenable to use by chevron.
}
\author{
\strong{Maintainer}: Benoit Falquet \email{benoit.falquet@roche.com}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
