% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{get_log}
\alias{get_log}
\alias{get_log.data.frame}
\alias{get_log.list}
\title{Get Log}
\usage{
get_log(data, incl)

\method{get_log}{data.frame}(data, incl = TRUE)

\method{get_log}{list}(data, incl = TRUE)
}
\arguments{
\item{data}{(\code{list} of \code{data.frame} or \code{data.frame}) filtered with \code{log_filter}.}

\item{incl}{(\code{flag}) should information about unfiltered \code{data.frame} be printed.}
}
\description{
Get Log
}
\examples{
data <- log_filter(iris, Sepal.Length >= 7)
get_log(data)

data <- log_filter(list(iris1 = iris, iris2 = iris), Sepal.Length >= 7, "iris1", character(0))
get_log(data)

}
