% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{h_expand_all_datasets}
\alias{h_expand_all_datasets}
\title{Propagate the rules for all datasets}
\usage{
h_expand_all_datasets(format_list, ls_datasets = NULL)
}
\arguments{
\item{ls_datasets}{(\code{character}) the name of all datasets in the object to reformat.}
}
\description{
Propagate the rules for all datasets
}
\examples{
\dontrun{
format_list <- list(
  adae = list(
    AEDECOD = "na_to_nca",
    AEBODSYS = "na_to_nca"
  ),
  all_datasets = list(AETOX = "na_to_nca")
)

h_expand_all_datasets(format_list, ls_datasets = c("adsl", "adae"))
}
}
\keyword{internal}
