% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{h_reformat_tab}
\alias{h_reformat_tab}
\title{Reformat a Variable in a Specific Column and Table}
\usage{
h_reformat_tab(db, tab, col, dic_map)
}
\arguments{
\item{db}{(\code{dm}) object input.}

\item{tab}{(\code{string}) the name of a table.}

\item{col}{(\code{string}) the name of a variable in a table.}

\item{dic_map}{(named \code{vector}) a dictionary with the mapping values, with the format \code{c(new = old)}.}
}
\value{
a \code{dm} object with re coded variables as factor.
}
\description{
Reformat a Variable in a Specific Column and Table
}
\note{
If \code{tab} is not a valid table name of the \code{db} object, the original object is returned. Similarly, if \code{col} is
not a valid column of the selected \code{tab} in the object, the original object is returned. This behavior is desirable
when a variable that exists in most but not all tables has to be re coded.
}
\examples{
library(dm)

df1 <- data.frame(
  "char" = c("a", "b", NA, "a", "k", "x"),
  "fact" = factor(c("f1", "f2", NA, NA, "f1", "f1")),
  "logi" = c(NA, FALSE, TRUE, NA, FALSE, NA)
)
df2 <- data.frame(
  "char" = c("a", "b", NA, "a", "k", "x"),
  "fact" = factor(c("f1", "f2", NA, NA, "f1", "f1")),
  "num" = 1:6
)

db <- dm(df1, df2)

dic_map <- setNames(c("A", "B", "Missing"), c("a", "b", NA))
res <- h_reformat_tab(db, "df1", "char", dic_map)
}
