% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_ws_to_na}
\alias{h_ws_to_na}
\title{Transforming Empty Strings and White Spaces to NAs}
\usage{
h_ws_to_na(x)
}
\arguments{
\item{x}{(\code{character} or \code{factor}) where empty of white space should be transformed to \code{NAs}.}
}
\value{
\code{character} or \code{factor} without explicit NA
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

SAS imports missing data as empty strings or white spaces. This helper function replaces the empty strings and white
space-only character and levels by \code{NAs}.
}
\examples{

char1 = c(" ", "    ", "a", "b", "", "")
h_ws_to_na(char1)

fact1 <- as.factor(char1)
h_ws_to_na(fact1)

}
