% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{list2rules}
\alias{list2rules}
\title{Read yaml File describing \code{rule}}
\usage{
list2rules(obj)
}
\arguments{
\item{file}{(\code{string}) of path to the rule yaml file.}
}
\description{
Read yaml File describing \code{rule}
}
\examples{
obj <- list(
  rule1 = list("X" = c("a", "b"), "Z" = "c"),
  rule2 = list(Missing = c(NA, ""))
)
list2rules(obj)

}
