% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{print_log}
\alias{print_log}
\alias{print_log.data.frame}
\alias{print_log.list}
\title{Print Log}
\usage{
print_log(data, ...)

\method{print_log}{data.frame}(data, incl = TRUE, ...)

\method{print_log}{list}(data, incl = TRUE, ...)
}
\description{
Print Log
}
\examples{
data <- log_filter(iris, Sepal.Length >= 7)
print_log(data)
data <- log_filter(list(iris1 = iris, iris2 = iris), Sepal.Length >= 7, "iris1", character(0))
print_log(data)
}
