% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{reformat}
\alias{reformat}
\title{Reformat Values}
\usage{
reformat(obj, format)
}
\arguments{
\item{obj}{object to reformat}

\item{format}{(\code{rule}) or (\code{list}) of \code{rules} depending on the class of obj}
}
\description{
Reformat Values
}
\details{
For character values, reformatting will only change those should be changed.
NA values are not changed, if the rule does not contain \code{NA_character_}.
Factors works similarly to characters.
For \code{dm} objects, see \code{apply_reformat} for more details.
\code{empty_rule} will do nothing to the data.
}
