% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{reformat}
\alias{reformat}
\alias{reformat.default}
\alias{reformat.character}
\alias{reformat.factor}
\alias{reformat.dm}
\alias{reformat.list}
\title{Reformat Values}
\usage{
reformat(obj, ...)

\method{reformat}{default}(obj, format, ...)

\method{reformat}{character}(obj, format, string_as_fct = TRUE, na_last = TRUE, ...)

\method{reformat}{factor}(obj, format, na_last = TRUE, ...)

\method{reformat}{dm}(obj, format, string_as_fct = TRUE, na_last = TRUE, ...)

\method{reformat}{list}(obj, format, string_as_fct = TRUE, na_last = TRUE, ...)
}
\arguments{
\item{obj}{object to reformat.}

\item{...}{not used. Only for compatibility between methods.}

\item{format}{(\code{list}) or (\code{list}) of \code{lists} depending on the class of obj.}

\item{string_as_fct}{(\code{flag}) whether the reformatted character object should be converted to factor.}

\item{na_last}{(\code{flag}) whether the level replacing \code{NA} should be last.}
}
\description{
Reformat Values
}
\note{
Values subject to reformatting but absent from the object do not lead to an error. The rest of the reformatting
process (for instance the conversion to factor  and the reformatting of factors levels if \code{string_as_fct = TRUE})
is still carried out.
}
\examples{

# Reformatting of character.
obj <- c("a", "b", "x", NA)
attr(obj, "label") <- "my label"
format <- list("A" = "a", "NN" = c(NA, ""))

reformat(obj, format, string_as_fct = FALSE)

# Reformatting of factor.
obj <- factor(c("a", "aa", "x", "NA", "b"), levels = c("x", "b", "aa", "a", "z"))
attr(obj, "label") <- "my label"
format <- list("A" = c("a", "aa"), "Missing" = c(NA, "x"), "Not Present" = "z")

reformat(obj, format)
reformat(obj, format, na_last = FALSE)

# Reformatting of dm.
library(dm)

df1 <- data.frame(
  var1 = c("a", "b", NA),
  var2 = factor(c("F1", "F2", NA))
)

df2 <- data.frame(
  var1 = c("x", NA, "y"),
  var2 = factor(c("F11", NA, "F22"))
)

db <- dm(df1 = df1, df2 = df2)

format <- list(
  df1 = list(
    var1 = list("X" = "x", "N" = NA)
  ),
  df2 = list(
    var1 = list(),
    var2 = list("f11" = "F11", "NN" = NA)
  )
)

reformat(db, format)

# Reformatting of list of data.frame.
df1 <- data.frame(
  var1 = c("a", "b", NA),
  var2 = factor(c("F1", "F2", NA))
)

df2 <- data.frame(
  var1 = c("x", NA, "y"),
  var2 = factor(c("F11", NA, "F22"))
)

db <- list(df1 = df1, df2 = df2)

format <- list(
  df1 = list(
    var1 = list("X" = "x", "N" = c(NA, ""))
  ),
  df2 = list(
    var1 = list(),
    var2 = list("f11" = "F11", "NN" = NA)
  )
)

reformat(db, format)
}
