% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{reformat}
\alias{reformat}
\alias{reformat.default}
\alias{reformat.logical}
\alias{reformat.character}
\alias{reformat.factor}
\alias{reformat.dm}
\alias{reformat.list}
\title{Reformat Values}
\usage{
reformat(obj, ...)

\method{reformat}{default}(obj, format, ...)

\method{reformat}{logical}(obj, format, bool_as_fct = FALSE, na_last = TRUE, ...)

\method{reformat}{character}(obj, format, string_as_fct = TRUE, na_last = TRUE, ...)

\method{reformat}{factor}(obj, format, na_last = TRUE, ...)

\method{reformat}{dm}(obj, format, string_as_fct = TRUE, na_last = TRUE, ...)

\method{reformat}{list}(obj, format, string_as_fct = TRUE, na_last = TRUE, ...)
}
\arguments{
\item{obj}{object to reformat.}

\item{...}{not used. Only for compatibility between methods.}

\item{format}{(\code{rule}) or (\code{list}) of \code{rule} depending on the class of obj.}

\item{na_last}{(\code{flag}) whether the level replacing \code{NA} should be last.}

\item{string_as_fct}{(\code{flag}) whether the reformatted character object should be converted to factor.}
}
\description{
Reformat Values
}
\note{
When the rule is empty rule or when values subject to reformatting are absent from the object, no error is
raised. The rest of the reformatting process (for instance the conversion to factor  and the reformatting of
factors levels if \code{string_as_fct = TRUE}) is still carried out.
}
\examples{

# Reformatting of character.
obj <- c(TRUE, FALSE, TRUE, NA)
attr(obj, "label") <- "my label"
format <- rule("<Missing>" = NA)

reformat(obj, format)
reformat(obj, format, bool_as_fct = TRUE)

# Reformatting of character.
obj <- c("a", "b", "x", NA)
attr(obj, "label") <- "my label"
format <- rule("A" = "a", "NN" = NA)

reformat(obj, format)

# Reformatting of factor.
obj <- factor(c("a", "aa", "b", "x", NA), levels = c("x", "b", "aa", "a", "z"))
attr(obj, "label") <- "my label"
format <- rule("A" = c("a", "aa"), "NN" = c(NA, "x"), "Not Present" = "z", "Not A level" = "P")

reformat(obj, format)
reformat(obj, format, na_last = TRUE)

# Reformatting of dm.
library(dm)

df1 <- data.frame(
  var1 = c("a", "b", NA),
  var2 = factor(c("F1", "F2", NA))
)

df2 <- data.frame(
  var1 = c("x", NA, "y"),
  var2 = factor(c("F11", NA, "F22"))
)

db <- dm(df1 = df1, df2 = df2)

format <- list(
  df1 = list(
    var1 = rule("X" = "x", "N" = NA)
  ),
  df2 = list(
    var1 = empty_rule,
    var2 = rule("f11" = "F11", "NN" = NA)
  )
)

reformat(db, format)

# Reformatting of list of data.frame.
df1 <- data.frame(
  var1 = c("a", "b", NA),
  var2 = factor(c("F1", "F2", NA))
)

df2 <- data.frame(
  var1 = c("x", NA, "y"),
  var2 = factor(c("F11", NA, "F22"))
)

db <- list(df1 = df1, df2 = df2)

format <- list(
  df1 = list(
    var1 = rule("X" = "x", "N" = NA)
  ),
  df2 = list(
    var1 = empty_rule,
    var2 = rule("f11" = "F11", "NN" = NA)
  )
)

reformat(db, format)
}
