% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_safe.R
\name{render_safe}
\alias{render_safe}
\title{Render whiskers safely}
\usage{
render_safe(x)
}
\arguments{
\item{x}{(\code{character}) input to be rendered safely.}
}
\value{
\code{character} with substituted placeholders.
}
\description{
Render whiskers safely
}
\note{
The strings enclosed in \code{{}} are substituted using the key-values pairs set with \code{add_whiskers}.
}
\examples{
render_safe("Name of {Patient_label}")
}
