% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{rule}
\alias{rule}
\title{Create rule based on mappings}
\usage{
rule(..., .lst = list(...))
}
\arguments{
\item{...}{Mapping pairs, the argument name is the transformed while
its values are original values.}

\item{.lst}{(\code{list}) of mapping.}
}
\description{
Create rule based on mappings
}
\note{
Conversion to \code{NA} is the last step of the remapping process.

Special mapping values include:
\itemize{
\item \code{.drop} (\code{flag}) whether to drop empty levels.
\item \code{.to_NA} (\code{character}) values that should be converted to \code{NA}.
}
}
\examples{
rule("X" = "x", "Y" = c("y", "z"))
rule("X" = "x", "Y" = c("y", "z"), .drop = TRUE, .to_NA = c("a", "b"))

}
