% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{rule}
\alias{rule}
\title{Create rule based on mappings}
\usage{
rule(..., .lst = list(...))
}
\arguments{
\item{...}{Mapping pairs, the argumen name is the transformed while
its values are original values.}

\item{.lst}{(\code{list}) of mapping.}
}
\description{
Create rule based on mappings
}
