% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_safe.R
\name{safe_transformer}
\alias{safe_transformer}
\title{Safe transformer}
\usage{
safe_transformer(text, envir)
}
\arguments{
\item{text}{(\code{string}) to be substituted.}

\item{envir}{(\code{environment}) containing key-value pairs describing the substitution to perform.}
}
\value{
\code{string} with substituted placeholders.
}
\description{
Safe transformer
}
\details{
Obtain content in global environment by default.
If not found, use the environment here. The function first looks for an exact match. If not found, it searches for a
match in lower case then apply to the result the same case as the original value.
}
\keyword{internal}
