% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{standard_remap}
\alias{standard_remap}
\title{Example of Remapping List}
\format{
a nest \code{list} with \code{rule}.
}
\source{
this object is based on the \code{yaml} file of the same name stored in \code{inst/extdata}.
}
\usage{
standard_remap
}
\description{
Example of Remapping List
}
\keyword{datasets}
